@@ -2,49 +2,34 @@
   "name": "couragescores",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
+    "@types/websocket": "^1.0.10",
+    "@vitejs/plugin-react": "^4.2.1",
     "bootstrap": "^5.1.3",
-    "http-proxy-middleware": "^2.0.6",
     "jest-junit": "^16.0.0",
-    "jquery": "^3.6.0",
-    "merge": "^2.1.1",
-    "oidc-client": "^1.11.5",
     "react-markdown": "^8.0.7",
     "react-router-bootstrap": "^0.26.1",
     "react-router-dom": "^6.3.0",
     "reactstrap": "^9.1.1",
     "remark-gfm": "^3.0.1",
-    "rimraf": "^3.0.2",
-    "web-vitals": "^2.1.4",
-    "workbox-background-sync": "^6.5.3",
-    "workbox-broadcast-update": "^6.5.3",
-    "workbox-cacheable-response": "^6.5.3",
-    "workbox-core": "^6.5.3",
-    "workbox-expiration": "^6.5.3",
-    "workbox-google-analytics": "^6.5.3",
-    "workbox-navigation-preload": "^6.5.3",
-    "workbox-precaching": "^6.5.3",
-    "workbox-range-requests": "^6.5.3",
-    "workbox-routing": "^6.5.3",
-    "workbox-strategies": "^6.5.3",
-    "workbox-streams": "^6.5.3"
+    "vite": "^5.0.12",
+    "vite-tsconfig-paths": "^4.2.3"
   },
   "devDependencies": {
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
-    "ajv": "^8.11.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.18.0",
+    "@types/jest": "^29.5.11",
+    "eslint": "^8.56.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.0",
-    "eslint-plugin-react": "^7.30.1",
-    "nan": "^2.16.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "typescript": "^4.7.4"
+    "ts-jest": "^29.1.2",
+    "typescript": "^4.7.4",
+    "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-mkcert": "^1.17.3"
   },
   "overrides": {
     "autoprefixer": "10.4.5"
@@ -54,26 +39,29 @@
     "nth-check": "^3.0.1"
   },
   "jest": {
+    "preset": "ts-jest/presets/js-with-ts",
+    "testEnvironment": "jest-environment-jsdom",
     "coverageReporters": [
       "lcov"
     ],
     "moduleNameMapper": {
       "next/router": ",<rootDir>/__mocks__/next/router.js",
-      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy",
-      "^.+\\.(jpg|jpeg|png|gif|webp|avif|svg)$": "<rootDir>/__mocks__/file-mock.js",
+      "\\.(css|less)$": "<rootDir>/src/__mocks__/style-mock.js",
+      "^.+\\.(jpg|jpeg|png|gif|webp|avif|svg)$": "<rootDir>/src/__mocks__/file-mock.js",
       "react-markdown": "<rootDir>/node_modules/react-markdown/react-markdown.min.js"
     },
     "collectCoverageFrom": [
-      "!src/api/**/*.js",
-      "!src/tests/**/*.js"
+      "!src/api/**/*.ts",
+      "!src/tests/**/*.ts",
+      "!coverage/**/*.js"
     ]
   },
   "scripts": {
-    "start": "rimraf ./build && react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src/",
+    "start": "vite",
+    "compile-ts": "tsc",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "jest --env=jsdom",
     "postpublish": "node ./src/whitelabel-publish.js"
   },
   "eslintConfig": {
